feat: Add GetFileSize function to file_manager
This commit adds the `GetFileSize` function to the `file_manager` class. This function retrieves the size of a file given its path. It uses the `stat` function to get the file's information, and if successful, updates the `size` parameter with the file size. If the `stat` function fails, an error message is logged and the `size` parameter is set to 0. This function enhances the functionality of the `file_manager` class by providing a convenient way to obtain the size of a file.

Refs: #123